Flattened model structure and removed The Folder structure of the models forlder was adapted to not contain a folder for every model, since we would like to move more into the direction of having one dynamics model which can be configurated by the config file. Furthermore, the plot_model_prediction was removed from the generate_dynamics_model.py file since it is only present in the simple quadrotor model but not the quadplane model.
