+++ Tools/parametric_model/src/models/__init__.py
@@ -4,8 +4,11 @@ __author__ = "Manuel Galliker"
 __maintainer__ = "Manuel Galliker"
 __license__ = "BSD 3"
 
-from . import simple_quadrotor_model
 from . import aerodynamic_models
 from . import rotor_models
 from . import model_plots
+from .dynamics_model import DynamicsModel
 from . import quad_plane_model
+from .quad_plane_model import QuadPlaneModel
+from . import simple_quadrotor_model
+from .simple_quadrotor_model import SimpleQuadRotorModel
