--- Tools/parametric_model/src/models/quad_plane_model/__init__.py
@@ -1,2 +1,0 @@
-from . import quad_plane_model
-from .quad_plane_model import QuadPlaneModel
