--- Tools/parametric_model/src/models/simple_quadrotor_model/__init__.py
@@ -1,2 +1,0 @@
-from . import simple_quadrotor_model
-from .simple_quadrotor_model import SimpleQuadRotorModel
